Outsourced Matrix classes into own files
