+++ src/Moserware/Skills/Numerics/Matrix/DiagonalMatrix.php
@@ -1,0 +1,31 @@
+<?php
+namespace Moserware\Numerics\Matrix;
+
+use Moserware\Numerics\Matrix;
+
+class DiagonalMatrix extends Matrix
+{
+    public function __construct(array $diagonalValues)
+    {
+        $diagonalCount = count($diagonalValues);
+        $rowCount = $diagonalCount;
+        $colCount = $rowCount;
+
+        parent::__construct($rowCount, $colCount);
+
+        for($currentRow = 0; $currentRow < $rowCount; $currentRow++)
+        {
+            for($currentCol = 0; $currentCol < $colCount; $currentCol++)
+            {
+                if($currentRow == $currentCol)
+                {
+                    $this->setValue($currentRow, $currentCol, $diagonalValues[$currentRow]);
+                }
+                else
+                {
+                    $this->setValue($currentRow, $currentCol, 0);
+                }
+            }
+        }
+    }
+}
